+++ pkg/datasources/redis/models.go
@@ -1,8 +1,15 @@
 package redis
 
+type TargetFormat string
+
+const (
+	TimeSeries TargetFormat = "time_series"
+	Heatmap    TargetFormat = "heatmap"
+)
+
 // Query is a single redis query (target)
 type Query struct {
-	Expr         string `json:"expr"`
-	Format       string `json:"format"`
-	LegendFormat string `json:"legendFormat"`
+	Expr         string       `json:"expr"`
+	Format       TargetFormat `json:"format"`
+	LegendFormat string       `json:"legendFormat"`
 }